+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Power switch entity that shows current power state using `is_powered_on()` and allows toggling power on/off
+- Upgrade `hdmimatrix` library to `v0.3.0`
+
 
 ## [0.2.0] - 2026-02-25
 ### Added